requirements.txt
accelerate==0.24.1
auth-token==0.0.3
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
customtkinter==0.3
darkdetect==0.7.1
diffusers==0.23.1
filelock==3.8.0
huggingface-hub==0.9.1
idna==3.4
importlib-metadata==4.12.0
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
tk==0.1.0
tokenizers==0.12.1
torch==1.13.0+cu116
torchaudio==0.13.0
torchvision==0.13.0+cu116
tqdm==4.64.1
transformers==4.35.2
typing_extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1